agents-layer/requirements.txt
python-dotenv>=1.0.0
httpx>=0.27.0
kubernetes>=30.1.0
PyYAML>=6.0.1
qdrant-client>=1.9.0
fastembed>=0.3.0
openai>=1.40.0
pydantic>=2.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
redis>=5.2.0
pytest>=8.0.0
langgraph>=0.6.0
langchain-openai>=0.0.0
